Rename "simplify.py" to "property_values.py"
Simplify is too vague; this file really pertains to notion property
values.

We should probably rename "converter.py" to "blocks.py" and split out
the rich-text related stuff.
